Update fee to not return concrete errors
